sac/backend/requirements.txt
# SAC Backend Dependencies

# Flask
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
gunicorn>=21.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Redis
redis>=5.0.0

# OpenAI & LangChain
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector Database (pgvector - uses existing PostgreSQL)
pgvector>=0.3.0
langchain-postgres>=0.0.9
psycopg[binary]>=3.1.0

# Data Validation
pydantic>=2.0.0

# Retry Logic
tenacity>=8.0.0

# Document Processing
markitdown>=0.1.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
